Choose syntax from file extension, not content.
