+++ lib/jekyll-sass-converter.rb
@@ -1,4 +1,5 @@
 require "jekyll-sass-converter/version"
+require "jekyll/converters/scss"
 require "jekyll/converters/sass"
 
 module JekyllSassConverter
