+++ lib/jekyll/converters/sass.rb
@@ -3,64 +3,16 @@ require 'jekyll/utils'
 
 module Jekyll
   module Converters
-    class Sass < Converter
+    class Sass < Scss
       safe true
       priority :low
 
       def matches(ext)
-        ext =~ /^\.s(a|c)ss$/i
+        ext =~ /^\.sass$/i
       end
 
-      def output_ext(ext)
-        ".css"
-      end
-
-      def jekyll_sass_configuration
-        options = @config["sass"] || {}
-        unless options["style"].nil?
-          options["style"] = options["style"].to_s.gsub(/\A:/, '').to_sym
-        end
-        options
-      end
-
-      def sass_build_configuration_options(overrides)
-        Jekyll::Utils.symbolize_hash_keys(
-          Jekyll::Utils.deep_merge_hashes(jekyll_sass_configuration, overrides)
-        )
-      end
-
-      def syntax_type_of_content(content)
-        if content.include?(";") || content.include?("{")
-          :scss
-        else
-          :sass
-        end
-      end
-
-      def sass_dir
-        return "_sass" if jekyll_sass_configuration["sass_dir"].to_s.empty?
-        jekyll_sass_configuration["sass_dir"]
-      end
-
-      def sass_dir_relative_to_site_source
-        # FIXME: Not Windows-safe. Can only change once Jekyll 2.0.0 is out
-        Jekyll.sanitized_path(@config["source"], sass_dir)
-      end
-
-      def allow_caching?
-        !@config["safe"]
-      end
-
-      def sass_configs(content = "")
-        sass_build_configuration_options({
-          "syntax" => syntax_type_of_content(content),
-          "cache"  => allow_caching?,
-          "load_paths" => [sass_dir_relative_to_site_source]
-        })
-      end
-
-      def convert(content)
-        ::Sass.compile(content, sass_configs(content))
+      def sass_syntax(content)
+        :sass
       end
     end
   end
